requirements.txt
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2023.4
gunicorn==20.1.0
dj-database-url==0.5.0
